+++ example/integration_test/src/tests/unit/unit_tests.dart
@@ -11,7 +11,7 @@ unitTests(WidgetTester tester) async {
 
   await testDeviceInfo();
 
-  // testHandleError(tester);
+  testHandleError(tester);
 
   testAppSettings(tester);